requirements.txt
absl-py==2.1.0
ale-py==0.8.1
alembic==1.14.0
astunparse==1.6.3
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.0
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.67.1
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h5py==3.12.1
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
keras==2.13.1
keyboard==0.13.5
kiwisolver==1.4.7
libclang==18.1.1
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.2
namex==0.0.8
networkx==3.4.2
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
optuna==4.1.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==4.25.5
psutil==6.1.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pynput==1.7.7
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scipy==1.14.1
seaborn==0.13.2
Shimmy==1.3.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==1.4.2
stable_baselines3==2.3.2
sympy==1.13.1
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tqdm==4.67.0
typing_extensions==4.5.0
tzdata==2024.2
ultralytics==8.3.28
ultralytics-thop==2.0.11
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.16.0
